+++ AzureCommunicationUI/sdk/AzureCommunicationUIChat/Sources/Presentation/SwiftUI/Chat/ChatViewComponents/ObservableScrollView.swift
@@ -1,0 +1,65 @@
+//
+//  Copyright (c) Microsoft Corporation. All rights reserved.
+//  Licensed under the MIT License.
+//
+
+import SwiftUI
+
+struct ScrollOffsetPreferenceKey: PreferenceKey {
+    static var defaultValue: CGFloat = .zero
+    static func reduce(value: inout CGFloat, nextValue: () -> CGFloat) {}
+}
+
+struct ScrollHeightPreferenceKey: PreferenceKey {
+    typealias Value = CGFloat
+    static var defaultValue: CGFloat {0}
+    static func reduce(value: inout CGFloat, nextValue: () -> CGFloat) {
+        value += nextValue()
+    }
+}
+
+struct ObservableScrollView<Content: View>: View {
+    let axes: Axis.Set
+    let showsIndicators: Bool
+    let offsetChanged: (CGFloat) -> Void
+    let heightChanged: (CGFloat) -> Void
+    let content: Content
+
+    init(
+        axes: Axis.Set = .vertical,
+        showsIndicators: Bool = true,
+        offsetChanged: @escaping (CGFloat) -> Void = { _ in },
+        heightChanged: @escaping (CGFloat) -> Void = { _ in },
+        @ViewBuilder content: () -> Content
+    ) {
+        self.axes = axes
+        self.showsIndicators = showsIndicators
+        self.offsetChanged = offsetChanged
+        self.heightChanged = heightChanged
+        self.content = content()
+    }
+
+    var body: some View {
+        SwiftUI.ScrollView(axes, showsIndicators: showsIndicators) {
+            GeometryReader { geometry in
+                Color.clear.preference(
+                    key: ScrollOffsetPreferenceKey.self,
+                    value: -geometry.frame(in: .named("scrollView")).origin.y
+                )
+            }
+            content
+            .background(
+                GeometryReader { geometry in
+                    Color.clear.preference(key: ScrollHeightPreferenceKey.self, value: geometry.size.height)
+                }
+            )
+        }
+        .background(
+            GeometryReader { geometry in
+                Color.clear.preference(key: ScrollHeightPreferenceKey.self, value: -geometry.size.height)
+            })
+        .coordinateSpace(name: "scrollView")
+        .onPreferenceChange(ScrollOffsetPreferenceKey.self, perform: offsetChanged)
+        .onPreferenceChange(ScrollHeightPreferenceKey.self, perform: heightChanged)
+    }
+}
